backend/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
asyncio-mqtt==0.16.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
dj-rest-auth==2.2.0
django-allauth==0.61.0
google-auth==2.25.0
google-auth-oauthlib==1.2.0
et_xmlfile==2.0.0
executing==2.2.0
feedparser==6.0.11
git-filter-repo==2.47.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
lxml==5.4.0
matplotlib-inline==0.1.7
msgpack==1.1.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.0
opcua==0.98.13
openpyxl==3.1.5
packaging==25.0
paho-mqtt==2.1.0
pandarallel==1.6.5
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
redis==6.2.0
requests==2.32.4
ruff==0.12.1
sec-api==1.0.32
service-identity==24.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
waitress==3.0.2
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
